+++ packages/use-ref-from/src/useRefFrom.ts
@@ -1,4 +1,7 @@
-import { useMemo, useRef, type RefObject } from 'react';
+import React, { type RefObject } from 'react';
+
+// Need to destructure instead of import for React < 16.14.0.
+const { useMemo, useRef } = React;
 
 export default function useRefFrom<T>(value: T): RefObject<T> & { get current(): T } {
   const ref = useRef<T>();