+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Import React using destructuring for Node.js test runner, by [@compulim](https://github.com/compulim) in PR [#75](https://github.com/compulim/use-ref-from/pull/75)
+
 ## [0.2.0] - 2025-12-24
 
 ### Changed